:recycle: Updating SEO component to match codebase conventions
